Moved ctx to common directory
