+++ frontend/components/account/WaitForVerificationComponent.tsx
@@ -1,4 +1,4 @@
-import { useSession } from '@/ctx';
+import { useSession } from '@/common/ctx';
 import { Button, Card, HelperText, Surface, Text, TextInput } from 'react-native-paper';
 import { useEffect, useState } from 'react';
 import { ensureString, fetchUser, isMobile, sleep } from '@/common/utils';
